refactor(sidebar): Timeline and CreateStory belongs to the same page
Motivation: reuse the same Story api
