+++ src/pages/Story/index.js
@@ -1,1 +1,2 @@
-export * from './Story'
+export * from './CreateStory'
+export * from './Timeline'
