--- src/pages/Timeline/index.js
@@ -1,1 +1,0 @@
-export * from './Timeline'
